requirements.txt
comet_ml
evaluate
librosa
munch
python-dotenv
PyYAML
torch
tqdm
whisper
